Task 5

+++ app/src/main/java/com/nix/summer/finaly/core/entity/Coffee.kt
@@ -1,7 +1,7 @@
 package com.nix.summer.finaly.core.entity
 
-enum class Coffee(val water: Int, val milk: Int, val coffeeBeans: Int, val money: Int) {
-    ESPRESSO(250, 0, 15, 4),
-    LATTE(350, 75, 20, 7),
-    CAPPUCCINO(200, 100, 12, 6);
+enum class Coffee(val water: Int, val milk: Int, val coffeeBeans: Int, val money: Float) {
+    ESPRESSO(250, 0, 15, 4f),
+    LATTE(350, 75, 20, 7f),
+    CAPPUCCINO(200, 100, 12, 6f);
 }
